move aggregateFragment to core module
